pyproject.toml
[project]
name = "bolojax"
dynamic = ["version"]
description = "Bolometric sensitivity calculator for CMB instruments"
authors = [{ name = "Brodi Elwood", email = "bdelwood@fas.harvard.edu" }]
requires-python = ">=3.10"
readme = "README.md"
license = "BSD-3-Clause"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
]
dependencies = [
    "numpy>=2",
    "scipy",
    "pyyaml",
    "pydantic>=2.7",
    "jax[cpu]",
    "jaxtyping",
    "equinox",
    "zodiax>=0.5",
    "am-python",
    "joblib",
    "pint>=0.24",
    "xarray",
]

[project.optional-dependencies]
gpu = ["jax[cuda13]"]

[project.scripts]
bolojax = "bolojax.cli:main"

[project.urls]
Homepage = "https://github.com/bdelwood/bolojax"
Repository = "https://github.com/bdelwood/bolojax"

[dependency-groups]
dev = [
    # Testing
    "pytest>=8",
    "coverage[toml]>=7",
    # Linting & type checking
    "ruff>=0.9",
    "pre-commit>=3",
    "pyrefly>=0.58",
    # Docs
    "sphinx>=7",
    "furo>=2024",
    "myst-parser>=3",
    "sphinx-autoapi>=3",
    "sphinx-math-dollar>=1.2",
    "sphinx-autobuild>=2024",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch]
version.source = "vcs"
version.vcs = "git"
version.raw-options.version_scheme = "no-guess-dev"
metadata.allow-direct-references = true
build.hooks.vcs.version-file = "src/bolojax/_version.py"

[tool.hatch.build.targets]
sdist = { include = ["src/bolojax"] }

[tool.hatch.build.targets.wheel]
include = ["src/bolojax"]

[tool.hatch.build.targets.wheel.sources]
"src/bolojax" = "bolojax"

[tool.pytest.ini_options]
minversion = "9.0"
addopts = ["-ra", "--showlocals"]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:astropy",
    # numpy divide-by-zero in Planck function at T=0
    "ignore:divide by zero:RuntimeWarning",
    "ignore:invalid value:RuntimeWarning",
    # subprocess leak from astropy FITS writing
    "ignore:subprocess:ResourceWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
log_level = "INFO"
testpaths = ["tests"]

[tool.coverage]
run.source = ["bolojax"]
run.branch = true
report.show_missing = true
report.exclude_also = ['\.\.\.', 'if typing.TYPE_CHECKING:']

[tool.pyright]
typeCheckingMode = "basic"

[tool.ruff]
target-version = "py310"
line-length = 88
show-fixes = true
exclude = [".venv"]

[tool.ruff.lint]
extend-select = [
    "ARG",  # flake8-unused-arguments
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "EM",   # flake8-errmsg
    "EXE",  # flake8-executable
    "FA",   # flake8-future-annotations
    "FURB", # refurb
    "I",    # isort
    "ICN",  # flake8-import-conventions
    "ISC",  # flake8-implicit-str-concat
    "NPY",  # numpy-specific rules
    "PERF", # perflint
    "PGH",  # pygrep-hooks
    "PIE",  # flake8-pie
    "PL",   # pylint
    "PT",   # flake8-pytest-style
    "PTH",  # flake8-use-pathlib
    "RET",  # flake8-return
    "RSE",  # flake8-raise
    "RUF",  # ruff-specific
    "SIM",  # flake8-simplify
    "T10",  # flake8-debugger
    "T20",  # flake8-print
    "TRY",  # tryceratops
    "UP",   # pyupgrade
    "YTT",  # flake8-2020
]
ignore = [
    "PLR09",   # too many arguments/branches/etc
    "PLR2004", # magic value in comparison (physics code)
    "N802",    # math variable naming
    "N803",    # math variable naming
    "N806",    # math variable naming
    "ISC001",  # conflicts with ruff formatter
    "PLC0414", # X as X re-exports in __init__.py
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"nb/**" = ["T20"]
"scripts/**" = ["T20"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.codespell]
skip = "*.npy,uv.lock,docs/_build,*.fits"
ignore-words-list = "emiss,boloR"

[tool.pyrefly]
project-excludes = ["**/.venv*"]
ignore-missing-imports = ["scipy.*"]

[tool.pyrefly.errors]
# Var pydantic annotation factory
# pyrefly can't resolve Annotated[T, Validator()]
invalid-annotation = false
# jax.typing.ArrayLike arithmetic always returns Array at runtime
bad-return = false
# ArrayLike <> Array <> ndarray interop across jax/numpy boundary
bad-argument-type = false
# Accessing attrs on T | None PrivateAttrs (always set before use via set_parent/set_camera)
missing-attribute = false
# Operations on T | None values (same runtime invariant)
unsupported-operation = false
# np.interp overload resolution/unpacking numpy scalars
not-iterable = false
no-matching-overload = false
# Pydantic discriminated union: subclasses narrow the Literal type field
bad-override-mutable-attribute = false
# jax.Array <> np.ndarray assignment across boundary
bad-assignment = false
